// src/app/models/empresa.model.ts
export class Empresa {
    idEmpresa: string;
    razaoSocialEmpresa: string;
    cnpjEmpresa: string;
    telefoneEmpresa: string;
    enderecoEmpresa: string;
    numeroEmpresa: string;
    complementoEmpresa: string;
    bairroEmpresa: string;
    cidadeEmpresa: string;
    estadoEmpresa: string;
    cepEmpresa: string;
    emailEmpresa: string;
}